Improve typing, cleanup and reorganize.
